Move the code from presenter to bottom sheet view
