+++ bottom-sheet-delegate/BottomSheet/BottomSheetView.swift
@@ -4,12 +4,28 @@
 
 import UIKit
 
-final class BottomSheetView: UIView {
-    static let handleHeight: CGFloat = 20
+public protocol BottomSheetViewDelegate: AnyObject {
+    func bottomSheetViewDidReachDismissArea(_ view: BottomSheetView)
+}
+
+public final class BottomSheetView: UIView {
+    // MARK: - Public properties
+    public weak var delegate: BottomSheetViewDelegate?
+
+    public var isDimViewHidden: Bool {
+        get { dimView.isHidden }
+        set { dimView.isHidden = newValue }
+    }
 
     // MARK: - Private properties
 
     private let contentView: UIView
+    private let preferredHeights: [CGFloat]
+    private var topConstraint: NSLayoutConstraint!
+    private var currentTargetOffset: CGFloat = 0
+    private var targetOffsets = [CGFloat]()
+    private lazy var panGesture = UIPanGestureRecognizer(target: self, action: #selector(handlePan(panGesture:)))
+    private lazy var springAnimator = SpringAnimator(dampingRatio: 0.8, frequencyResponse: 0.4)
 
     private lazy var handleView: UIView = {
         let view = UIView(frame: .zero)
@@ -19,25 +35,90 @@ final class BottomSheetView: UIView {
         return view
     }()
 
+    private lazy var dimView: UIView = {
+        let view = UIView(frame: .zero)
+        view.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        view.backgroundColor = UIColor(white: 0, alpha: 0.4)
+        view.isHidden = true
+        view.alpha = 0
+        return view
+    }()
+
     // MARK: - Init
 
-    init(contentView: UIView) {
+    public init(contentView: UIView, preferredHeights: [CGFloat]) {
         self.contentView = contentView
+        self.preferredHeights = preferredHeights
         super.init(frame: .zero)
         setup()
     }
 
-    required init?(coder: NSCoder) {
+    public required init?(coder: NSCoder) {
         fatalError("Not implemented")
     }
 
     // MARK: - Overrides
 
-    override func layoutSubviews() {
+    public override func layoutSubviews() {
         super.layoutSubviews()
         // Make shadow to be on top
         let rect = CGRect(x: 0, y: 0, width: bounds.width, height: 30)
         layer.shadowPath = UIBezierPath(rect: rect).cgPath
+    }
+
+    // MARK: - Public API
+
+    public func present(in superview: UIView, completion: ((Bool) -> Void)? = nil) {
+        superview.addSubview(dimView)
+        superview.addSubview(self)
+
+        translatesAutoresizingMaskIntoConstraints = false
+
+        updateTargetOffsets()
+        dimView.frame = superview.bounds
+        topConstraint = topAnchor.constraint(equalTo: superview.topAnchor, constant: superview.frame.height)
+
+        springAnimator.addAnimation { [weak self] position in
+            self?.topConstraint.constant = position.y
+            self?.updateDimViewAlpha(for: position.y)
+        }
+
+        springAnimator.addCompletion { didComplete in completion?(didComplete) }
+
+        NSLayoutConstraint.activate([
+            topConstraint,
+            bottomAnchor.constraint(greaterThanOrEqualTo: superview.bottomAnchor),
+            leadingAnchor.constraint(equalTo: superview.leadingAnchor),
+            trailingAnchor.constraint(equalTo: superview.trailingAnchor)
+        ])
+
+        superview.layoutIfNeeded()
+        addGestureRecognizer(panGesture)
+        
+        animate(to: targetOffsets.last ?? 0)
+    }
+
+    public func hide(completion: ((Bool) -> Void)? = nil) {
+        springAnimator.addCompletion { [weak self] didComplete in
+            if didComplete {
+                self?.dimView.removeFromSuperview()
+                self?.removeFromSuperview()
+            }
+
+            completion?(didComplete)
+        }
+
+        animate(to: superview?.frame.height ?? 0)
+    }
+
+    public func reset() {
+        updateTargetOffsets()
+        animate(to: targetOffsets.last ?? 0)
+    }
+
+    public func transition<T: RawRepresentable>(to height: T) where T.RawValue == CGFloat {
+        guard let offset = offset(from: height.rawValue) else { return }
+        animate(to: offset)
     }
 
     // MARK: - Setup
@@ -72,4 +153,107 @@ final class BottomSheetView: UIView {
             contentView.bottomAnchor.constraint(lessThanOrEqualTo: bottomAnchor)
         ])
     }
+
+    // MARK: - Animations
+
+    private func animate(to constant: CGFloat) {
+        if targetOffsets.contains(constant) {
+            currentTargetOffset = constant
+        }
+
+        springAnimator.fromPosition = CGPoint(x: 0, y: topConstraint.constant)
+        springAnimator.toPosition = CGPoint(x: 0, y: constant)
+        springAnimator.initialVelocity = .zero
+        springAnimator.startAnimation()
+    }
+
+    private func updateDimViewAlpha(for offset: CGFloat) {
+        if let maxOffset = targetOffsets.last {
+            dimView.alpha = min(1, maxOffset / offset)
+        }
+    }
+
+    // MARK: - UIPanGestureRecognizer
+
+    @objc private func handlePan(panGesture: UIPanGestureRecognizer) {
+        let state = translationState(for: panGesture)
+
+        switch panGesture.state {
+        case .began:
+            springAnimator.pauseAnimation()
+        case .ended, .cancelled, .failed:
+            animate(to: state.targetOffset)
+            if state.isDismissible {
+                delegate?.bottomSheetViewDidReachDismissArea(self)
+            }
+        default:
+            break
+        }
+
+        topConstraint.constant = state.nextOffset
+        updateDimViewAlpha(for: state.nextOffset)
+        panGesture.setTranslation(.zero, in: superview)
+    }
+
+    // MARK: - Offset calculation
+
+    private func translationState(for panGesture: UIPanGestureRecognizer) -> TranslationState {
+        let currentArea = currentTargetOffset - .translationThreshold ... currentTargetOffset + .translationThreshold
+        let currentConstant = topConstraint.constant
+        let translation = panGesture.translation(in: superview)
+        let dragConstant = topConstraint.constant + translation.y
+
+        if currentArea.contains(dragConstant) {
+            return TranslationState(nextOffset: dragConstant, targetOffset: currentTargetOffset, isDismissible: false)
+        } else if dragConstant < currentTargetOffset {
+            let targetOffset = targetOffsets.first(where: { $0 < dragConstant })
+            return TranslationState(
+                nextOffset: targetOffset == nil ? currentConstant : dragConstant,
+                targetOffset: targetOffset ?? currentTargetOffset,
+                isDismissible: false
+            )
+        } else {
+            let targetOffset = targetOffsets.first(where: { $0 > dragConstant })
+            return TranslationState(
+                nextOffset: dragConstant,
+                targetOffset: targetOffset ?? currentTargetOffset,
+                isDismissible: targetOffset == nil
+            )
+        }
+    }
+
+    private func updateTargetOffsets() {
+        targetOffsets = preferredHeights.compactMap(offset(from:)).sorted()
+    }
+
+    private func offset(from height: CGFloat) -> CGFloat? {
+        guard let superview = superview else { return nil }
+
+        func makeTargetHeight() -> CGFloat {
+            if height == .bottomSheetAutomatic {
+                let size = contentView.systemLayoutSizeFitting(
+                    superview.frame.size,
+                    withHorizontalFittingPriority: .required,
+                    verticalFittingPriority: .fittingSizeLevel
+                )
+                return size.height
+            } else {
+                return height
+            }
+        }
+
+        let handleHeight: CGFloat = 20
+        let targetHeight = makeTargetHeight() + handleHeight
+        let minOffset: CGFloat = .translationThreshold
+
+        return max(superview.frame.height - max(targetHeight, minOffset), minOffset)
+    }
+}
+
+// MARK: - Private types
+
+private struct TranslationState {
+    let nextOffset: CGFloat
+    let targetOffset: CGFloat
+    let isDismissible: Bool
 }
